5. Create a class Main having a main method that does the following:
a. It creates a new Employee object employee (you can invent your own
name, hireday, salary, etc., to be used in the constructor)
b. Then it creates a checking account, savings account and retirement
account for employee, each with a starting balance of $300.
c. Then it prints to the console the account data for each of these
accounts (making use of the toString() method that has been provided in
Account)